Unify the make_hard_link function
